Added Timer class, added respawn and shield animations

+++ src/Game/Game.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <glm/vec2.hpp>
-#include <memory>
 #include <array>
+#include <memory>
 
 class Tank;
 class Level;
@@ -13,7 +13,7 @@ public:
     ~Game();
 
     void render();
-    void update(const uint64_t delta);
+    void update(const double delta);
     void setKey(const int key, const int action);
     bool init();
     size_t getCurrentLevelWidth() const;
@@ -29,7 +29,6 @@ private:
 
     glm::ivec2 m_windowSize;
     EGameState m_eCurrentGameState;
-
     std::unique_ptr<Tank> m_pTank;
     std::unique_ptr<Level> m_pLevel;
 };
